+++ crates/matrix-sdk-sqlite/src/utils.rs
@@ -13,6 +13,8 @@
 // limitations under the License.
 
 use core::fmt;
+#[cfg(target_family = "wasm")]
+use std::convert::Infallible;
 use std::{
     borrow::{Borrow, Cow},
     cmp::min,
@@ -21,8 +23,10 @@ use std::{
 };
 
 use async_trait::async_trait;
+#[cfg(not(target_family = "wasm"))]
 use deadpool_sync::InteractError;
 use itertools::Itertools;
+use matrix_sdk_base::SendOutsideWasm;
 use matrix_sdk_store_encryption::StoreCipher;
 use ruma::{OwnedEventId, OwnedRoomId, serde::Raw, time::SystemTime};
 use rusqlite::{OptionalExtension, Params, Row, Statement, Transaction, limits::Limit};
@@ -65,55 +69,62 @@ impl rusqlite::ToSql for Key {
     }
 }
 
-#[async_trait]
+#[cfg_attr(target_family = "wasm", async_trait(?Send))]
+#[cfg_attr(not(target_family = "wasm"), async_trait)]
 pub(crate) trait SqliteAsyncConnExt {
     async fn execute<P>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
     ) -> rusqlite::Result<usize>
     where
-        P: Params + Send + 'static;
+        P: Params + SendOutsideWasm + 'static;
 
-    async fn execute_batch(&self, sql: impl AsRef<str> + Send + 'static) -> rusqlite::Result<()>;
+    async fn execute_batch(
+        &self,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
+    ) -> rusqlite::Result<()>;
 
+    #[allow(dead_code)]
     async fn prepare<T, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         f: F,
     ) -> rusqlite::Result<T>
     where
-        T: Send + 'static,
-        F: FnOnce(Statement<'_>) -> rusqlite::Result<T> + Send + 'static;
+        T: SendOutsideWasm + 'static,
+        F: FnOnce(Statement<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static;
 
     async fn query_row<T, P, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
         f: F,
     ) -> rusqlite::Result<T>
     where
-        T: Send + 'static,
-        P: Params + Send + 'static,
-        F: FnOnce(&Row<'_>) -> rusqlite::Result<T> + Send + 'static;
+        T: SendOutsideWasm + 'static,
+        P: Params + SendOutsideWasm + 'static,
+        F: FnOnce(&Row<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static;
 
+    #[allow(dead_code)]
     async fn query_many<T, P, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
         f: F,
     ) -> rusqlite::Result<Vec<T>>
     where
-        T: Send + 'static,
-        P: Params + Send + 'static,
-        F: FnMut(&Row<'_>) -> rusqlite::Result<T> + Send + 'static;
+        T: SendOutsideWasm + 'static,
+        P: Params + SendOutsideWasm + 'static,
+        F: FnMut(&Row<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static;
 
     async fn with_transaction<T, E, F>(&self, f: F) -> Result<T, E>
     where
-        T: Send + 'static,
-        E: From<rusqlite::Error> + Send + 'static,
-        F: FnOnce(&Transaction<'_>) -> Result<T, E> + Send + 'static;
+        T: SendOutsideWasm + 'static,
+        E: From<rusqlite::Error> + SendOutsideWasm + 'static,
+        F: FnOnce(&Transaction<'_>) -> Result<T, E> + SendOutsideWasm + 'static;
 
+    #[allow(dead_code)]
     async fn chunk_large_query_over<Query, Res>(
         &self,
         mut keys_to_chunk: Vec<Key>,
@@ -121,8 +132,8 @@ pub(crate) trait SqliteAsyncConnExt {
         do_query: Query,
     ) -> Result<Vec<Res>>
     where
-        Res: Send + 'static,
-        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + Send + 'static;
+        Res: SendOutsideWasm + 'static,
+        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + SendOutsideWasm + 'static;
 
     /// Apply the [`RuntimeConfig`].
     ///
@@ -240,22 +251,26 @@ pub(crate) trait SqliteAsyncConnExt {
     }
 }
 
-#[async_trait]
+#[cfg_attr(target_family = "wasm", async_trait(?Send))]
+#[cfg_attr(not(target_family = "wasm"), async_trait)]
 impl SqliteAsyncConnExt for SqliteAsyncConn {
     async fn execute<P>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
     ) -> rusqlite::Result<usize>
     where
-        P: Params + Send + 'static,
+        P: Params + SendOutsideWasm + 'static,
     {
         self.interact(move |conn| conn.execute(sql.as_ref(), params))
             .await
             .map_err(map_interact_err)?
     }
 
-    async fn execute_batch(&self, sql: impl AsRef<str> + Send + 'static) -> rusqlite::Result<()> {
+    async fn execute_batch(
+        &self,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
+    ) -> rusqlite::Result<()> {
         self.interact(move |conn| conn.execute_batch(sql.as_ref()))
             .await
             .map_err(map_interact_err)?
@@ -263,26 +278,26 @@ impl SqliteAsyncConnExt for SqliteAsyncConn {
 
     async fn prepare<T, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         f: F,
     ) -> rusqlite::Result<T>
     where
-        T: Send + 'static,
-        F: FnOnce(Statement<'_>) -> rusqlite::Result<T> + Send + 'static,
+        T: SendOutsideWasm + 'static,
+        F: FnOnce(Statement<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static,
     {
         self.interact(move |conn| f(conn.prepare(sql.as_ref())?)).await.map_err(map_interact_err)?
     }
 
     async fn query_row<T, P, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
         f: F,
     ) -> rusqlite::Result<T>
     where
-        T: Send + 'static,
-        P: Params + Send + 'static,
-        F: FnOnce(&Row<'_>) -> rusqlite::Result<T> + Send + 'static,
+        T: SendOutsideWasm + 'static,
+        P: Params + SendOutsideWasm + 'static,
+        F: FnOnce(&Row<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static,
     {
         self.interact(move |conn| conn.query_row(sql.as_ref(), params, f))
             .await
@@ -291,14 +306,14 @@ impl SqliteAsyncConnExt for SqliteAsyncConn {
 
     async fn query_many<T, P, F>(
         &self,
-        sql: impl AsRef<str> + Send + 'static,
+        sql: impl AsRef<str> + SendOutsideWasm + 'static,
         params: P,
         f: F,
     ) -> rusqlite::Result<Vec<T>>
     where
-        T: Send + 'static,
-        P: Params + Send + 'static,
-        F: FnMut(&Row<'_>) -> rusqlite::Result<T> + Send + 'static,
+        T: SendOutsideWasm + 'static,
+        P: Params + SendOutsideWasm + 'static,
+        F: FnMut(&Row<'_>) -> rusqlite::Result<T> + SendOutsideWasm + 'static,
     {
         self.interact(move |conn| {
             let mut stmt = conn.prepare(sql.as_ref())?;
@@ -310,9 +325,9 @@ impl SqliteAsyncConnExt for SqliteAsyncConn {
 
     async fn with_transaction<T, E, F>(&self, f: F) -> Result<T, E>
     where
-        T: Send + 'static,
-        E: From<rusqlite::Error> + Send + 'static,
-        F: FnOnce(&Transaction<'_>) -> Result<T, E> + Send + 'static,
+        T: SendOutsideWasm + 'static,
+        E: From<rusqlite::Error> + SendOutsideWasm + 'static,
+        F: FnOnce(&Transaction<'_>) -> Result<T, E> + SendOutsideWasm + 'static,
     {
         self.interact(move |conn| {
             let txn = conn.transaction()?;
@@ -338,8 +353,8 @@ impl SqliteAsyncConnExt for SqliteAsyncConn {
         do_query: Query,
     ) -> Result<Vec<Res>>
     where
-        Res: Send + 'static,
-        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + Send + 'static,
+        Res: SendOutsideWasm + 'static,
+        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + SendOutsideWasm + 'static,
     {
         self.with_transaction(move |txn| {
             txn.chunk_large_query_over(keys_to_chunk, result_capacity, do_query)
@@ -348,6 +363,7 @@ impl SqliteAsyncConnExt for SqliteAsyncConn {
     }
 }
 
+#[cfg(not(target_family = "wasm"))]
 /// Map an [`InteractError`] into a [`rusqlite::Error`].
 ///
 /// An [`InteractError::Panic`] will panic. An [`InteractError::Cancelled`] will
@@ -363,6 +379,16 @@ fn map_interact_err(error: InteractError) -> rusqlite::Error {
     }
 }
 
+#[cfg(target_family = "wasm")]
+/// An unreachable function to avoid having to put conditional compilation
+/// everywhere we want to use [`ConnectionWrapper::interact()`].
+///
+/// This function should not be reachable under normal circumstance since
+/// [`ConnectionWrapper::interact()`] return [`Infallible`] as an error.
+fn map_interact_err(_error: Infallible) -> rusqlite::Error {
+    unreachable!()
+}
+
 pub(crate) trait SqliteTransactionExt {
     fn chunk_large_query_over<Key, Query, Res>(
         &self,
@@ -371,8 +397,8 @@ pub(crate) trait SqliteTransactionExt {
         do_query: Query,
     ) -> Result<Vec<Res>>
     where
-        Res: Send + 'static,
-        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + Send + 'static;
+        Res: SendOutsideWasm + 'static,
+        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + SendOutsideWasm + 'static;
 }
 
 impl SqliteTransactionExt for Transaction<'_> {
@@ -383,8 +409,8 @@ impl SqliteTransactionExt for Transaction<'_> {
         do_query: Query,
     ) -> Result<Vec<Res>>
     where
-        Res: Send + 'static,
-        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + Send + 'static,
+        Res: SendOutsideWasm + 'static,
+        Query: Fn(&Transaction<'_>, Vec<Key>) -> Result<Vec<Res>> + SendOutsideWasm + 'static,
     {
         // Divide by 2 to allow space for more static parameters (not part of
         // `keys_to_chunk`).
@@ -434,14 +460,16 @@ pub(crate) trait SqliteKeyValueStoreConnExt {
     /// Store the given value for the given key.
     fn set_kv(&self, key: &str, value: &[u8]) -> rusqlite::Result<()>;
 
+    #[allow(dead_code)]
     /// Store the given value for the given key by serializing it.
-    fn set_serialized_kv<T: Serialize + Send>(&self, key: &str, value: T) -> Result<()> {
+    fn set_serialized_kv<T: Serialize + SendOutsideWasm>(&self, key: &str, value: T) -> Result<()> {
         let serialized_value = rmp_serde::to_vec_named(&value)?;
         self.set_kv(key, &serialized_value)?;
 
         Ok(())
     }
 
+    #[allow(dead_code)]
     /// Removes the current key and value if exists.
     fn clear_kv(&self, key: &str) -> rusqlite::Result<()>;
 
@@ -477,7 +505,8 @@ impl SqliteKeyValueStoreConnExt for rusqlite::Connection {
 ///     "value" BLOB NOT NULL
 /// );
 /// ```
-#[async_trait]
+#[cfg_attr(target_family = "wasm", async_trait(?Send))]
+#[cfg_attr(not(target_family = "wasm"), async_trait)]
 pub(crate) trait SqliteKeyValueStoreAsyncConnExt: SqliteAsyncConnExt {
     /// Whether the `kv` table exists in this database.
     async fn kv_table_exists(&self) -> rusqlite::Result<bool> {
@@ -497,6 +526,7 @@ pub(crate) trait SqliteKeyValueStoreAsyncConnExt: SqliteAsyncConnExt {
             .optional()
     }
 
+    #[allow(dead_code)]
     /// Get the stored serialized value for the given key.
     async fn get_serialized_kv<T: DeserializeOwned>(&self, key: &str) -> Result<Option<T>> {
         let Some(bytes) = self.get_kv(key).await? else {
@@ -509,13 +539,15 @@ pub(crate) trait SqliteKeyValueStoreAsyncConnExt: SqliteAsyncConnExt {
     /// Store the given value for the given key.
     async fn set_kv(&self, key: &str, value: Vec<u8>) -> rusqlite::Result<()>;
 
+    #[allow(dead_code)]
     /// Store the given value for the given key by serializing it.
-    async fn set_serialized_kv<T: Serialize + Send + 'static>(
+    async fn set_serialized_kv<T: Serialize + SendOutsideWasm + 'static>(
         &self,
         key: &str,
         value: T,
     ) -> Result<()>;
 
+    #[allow(dead_code)]
     /// Clears the given value for the given key.
     async fn clear_kv(&self, key: &str) -> rusqlite::Result<()>;
 
@@ -569,7 +601,8 @@ pub(crate) trait SqliteKeyValueStoreAsyncConnExt: SqliteAsyncConnExt {
     }
 }
 
-#[async_trait]
+#[cfg_attr(target_family = "wasm", async_trait(?Send))]
+#[cfg_attr(not(target_family = "wasm"), async_trait)]
 impl SqliteKeyValueStoreAsyncConnExt for SqliteAsyncConn {
     async fn set_kv(&self, key: &str, value: Vec<u8>) -> rusqlite::Result<()> {
         let key = key.to_owned();
@@ -578,7 +611,7 @@ impl SqliteKeyValueStoreAsyncConnExt for SqliteAsyncConn {
         Ok(())
     }
 
-    async fn set_serialized_kv<T: Serialize + Send + 'static>(
+    async fn set_serialized_kv<T: Serialize + SendOutsideWasm + 'static>(
         &self,
         key: &str,
         value: T,
@@ -604,6 +637,7 @@ pub(crate) fn repeat_vars(count: usize) -> impl fmt::Display {
     iter::repeat_n("?", count).format(",")
 }
 
+#[allow(dead_code)]
 /// Convert the given `SystemTime` to a timestamp, as the number of seconds
 /// since Unix Epoch.
 ///
@@ -660,21 +694,25 @@ pub(crate) trait EncryptableStore {
         }
     }
 
+    #[allow(dead_code)]
     fn serialize_value(&self, value: &impl Serialize) -> Result<Vec<u8>> {
         let serialized = rmp_serde::to_vec_named(value)?;
         self.encode_value(serialized)
     }
 
+    #[allow(dead_code)]
     fn deserialize_value<T: DeserializeOwned>(&self, value: &[u8]) -> Result<T> {
         let decoded = self.decode_value(value)?;
         Ok(rmp_serde::from_slice(&decoded)?)
     }
 
+    #[allow(dead_code)]
     fn serialize_json(&self, value: &impl Serialize) -> Result<Vec<u8>> {
         let serialized = serde_json::to_vec(value)?;
         self.encode_value(serialized)
     }
 
+    #[allow(dead_code)]
     fn deserialize_json<T: DeserializeOwned>(&self, data: &[u8]) -> Result<T> {
         let decoded = self.decode_value(data)?;